Fix: minor bugs related to swipe function
