+++ src/components/Navbar/Navbar.tsx
@@ -4,28 +4,30 @@ import { NavItem, NavItemProps } from "./NavItem";
 import "./Navbar.css";
 
 export const Navbar = () => {
-  const { toggleCardExpand, toggleFilter, isCardExpanded } = useAppContext();
-  const { handleNewJoke, copyJokeToClipboard, currentJoke } = useJokesContext();
+  const { toggleCardExpand, toggleCategorySelector, isCardExpanded } =
+    useAppContext();
+
+  const { removeTopJoke } = useJokesContext();
 
   const navItems: NavItemProps[] = [
     {
       type: "button",
-      onClick: toggleFilter,
+      onClick: toggleCategorySelector,
       icon: "SlidersHorizontal",
     },
     { type: "button", onClick: toggleCardExpand, icon: "CircleHelp" },
     {
       type: "button",
       onClick: () => {
-        if (currentJoke) {
-          copyJokeToClipboard(currentJoke, isCardExpanded);
-        }
+        // if (currentJoke) {
+        //   copyJokeToClipboard(currentJoke, isCardExpanded);
+        // }
       },
       icon: "Copy",
     },
     {
       type: "button",
-      onClick: handleNewJoke,
+      onClick: removeTopJoke,
       icon: "CircleArrowRight",
     },
   ];
